send init order request only on form submit
